+++ Log/src/main/java/de/cubeisland/cubeengine/log/logger/InteractionLogger.java
@@ -47,14 +47,14 @@ public class InteractionLogger extends Logger<InteractionLogger.InteractionConfi
                     if ((block.getType().equals(Material.WOODEN_DOOR) && this.config.logDoor) || (block.getType().equals(Material.TRAP_DOOR) && this.config.logTrapDoor)
                             || (block.getType().equals(Material.FENCE_GATE) && this.config.logfenceGate))
                     {
-                        this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                        this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                                 block.getType(), ((Openable) blockData).isOpen() ? 1 : 0);
                     }
                     return;
                 case LEVER:
                     if (this.config.logLever)
                     {
-                        this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                        this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                                 block.getType(), ((Lever) blockData).isPowered() ? 1 : 0);
                     }
                     return;
@@ -62,21 +62,21 @@ public class InteractionLogger extends Logger<InteractionLogger.InteractionConfi
                 case WOOD_BUTTON:
                     if (this.config.logButtons)
                     {
-                        this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                        this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                                 block.getType(), null);
                     }
                     return;
                 case CAKE_BLOCK: // data: remaining slices
                     if (this.config.logCake)
                     {
-                        this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                        this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                                 block.getType(), ((Cake) blockData).getSlicesRemaining());
                     }
                     return;
                 case NOTE_BLOCK:
                     if (this.config.logNoteBlock)
                     {
-                        this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                        this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                                 block.getType(), (int) ((NoteBlock) blockData).getRawNote());
                     }
                     return;
@@ -84,7 +84,7 @@ public class InteractionLogger extends Logger<InteractionLogger.InteractionConfi
                 case DIODE_BLOCK_ON:
                     if (this.config.logDiode)
                     {
-                        this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                        this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                                 block.getType(), ((Diode) blockData).getDelay());
                     }
             }
@@ -96,7 +96,7 @@ public class InteractionLogger extends Logger<InteractionLogger.InteractionConfi
             {
                 case WOOD_PLATE:
                 case STONE_PLATE:
-                    this.module.getLogManager().logInteractLog(user.key, user.getLocation(this.helper),
+                    this.module.getLogManager().logInteractLog(user.key.intValue(), user.getLocation(this.helper),
                             block.getType(), null);
             }
         }
